backend/pyproject.toml
[project]
name = "sorooshx-backend"
version = "1.0.0"
description = "SorooshX Exchange Backend - Django REST API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "SorooshX Team", email = "dev@sorooshx.com" }]

dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.6.0",
    "django-environ>=0.11.2",
    "dj-database-url>=2.3.0",
    "psycopg[binary]>=3.2.0",
    "redis>=5.2.0",
    "channels>=4.2.0",
    "channels-redis>=4.2.0",
    "gunicorn>=23.0.0",
    "uvicorn[standard]>=0.32.0",
    "whitenoise>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "django-stubs>=5.1.0",
    "djangorestframework-stubs>=3.15.0",
    "pre-commit>=4.0.0",
    "ipython>=8.29.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "DJ",  # flake8-django
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
